Update Actions to require class-level name.
Actions which specified a name during instantiation (or after)
caused a subtle bug with instance caching for tables, causing
actions to overwrite each other in the cache. A longer-term
solution would be to do instance caching based on class rather
than name, but that requires an invasive change since "name"
is used to construct actions.

Change-Id: I9ad29c8abe59fbb595be6bfaa185de238dcb5314